+++ test/bootstrap/node.js
@@ -1,5 +1,7 @@
-var chai = require('chai')
-  , passport = require('chai-passport-strategy');
+const chai = require('chai');
+
+
+const passport = require('chai-passport-strategy');
 
 chai.use(passport);