+++ test/oauth2.state.session.test.js
@@ -1,17 +1,13 @@
-var OAuth2Strategy = require('../lib/strategy')
-  , AuthorizationError = require('../lib/errors/authorizationerror')
-  , TokenError = require('../lib/errors/tokenerror')
-  , InternalOAuthError = require('../lib/errors/internaloautherror')
-  , chai = require('chai')
-  , uri = require('url');
+/* global describe, it, before, expect */
+/* eslint-disable no-unused-expressions, consistent-return */
+const chai = require('chai');
+const uri = require('url');
+const OAuth2Strategy = require('../lib/strategy');
 
-
-describe('OAuth2Strategy', function() {
-  
-  describe('using default session state store', function() {
-    
-    describe('issuing authorization request', function() {
-      var strategy = new OAuth2Strategy({
+describe('OAuth2Strategy', () => {
+  describe('using default session state store', () => {
+    describe('issuing authorization request', () => {
+      const strategy = new OAuth2Strategy({
         authorizationURL: 'https://www.example.com/oauth2/authorize',
         tokenURL: 'https://www.example.com/oauth2/token',
         clientID: 'ABC123',
@@ -19,48 +15,50 @@ describe('OAuth2Strategy', function() {
         callbackURL: 'https://www.example.net/auth/example/callback',
         state: true
       },
-      function(accessToken, refreshToken, profile, done) {});
-      
-      
-      describe('that redirects to service provider', function() {
-        var request, url;
-  
-        before(function(done) {
+      (() => {}));
+
+
+      describe('that redirects to service provider', () => {
+        let request; let
+          url;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .redirect(function(u) {
+            .redirect((u) => {
               url = u;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
               req.session = {};
             })
             .authenticate();
         });
-  
-        it('should be redirected', function() {
-          var u = uri.parse(url, true);
+
+        it('should be redirected', () => {
+          const u = uri.parse(url, true);
           expect(u.query.state).to.have.length(32);
         });
-      
-        it('should save state in session', function() {
-          var u = uri.parse(url, true);
-        
+
+        it('should save state in session', () => {
+          const u = uri.parse(url, true);
+
           expect(request.session['oauth2:www.example.com'].state).to.have.length(32);
           expect(request.session['oauth2:www.example.com'].state).to.equal(u.query.state);
         });
       }); // that redirects to service provider
-      
-      describe('that redirects to service provider with other data in session', function() {
-        var request, url;
-  
-        before(function(done) {
+
+      describe('that redirects to service provider with other data in session', () => {
+        let request; let
+          url;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .redirect(function(u) {
+            .redirect((u) => {
               url = u;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
               req.session = {};
               req.session['oauth2:www.example.com'] = {};
@@ -68,49 +66,47 @@ describe('OAuth2Strategy', function() {
             })
             .authenticate();
         });
-  
-        it('should be redirected', function() {
-          var u = uri.parse(url, true);
+
+        it('should be redirected', () => {
+          const u = uri.parse(url, true);
           expect(u.query.state).to.have.length(32);
         });
-      
-        it('should save state in session', function() {
-          var u = uri.parse(url, true);
-        
+
+        it('should save state in session', () => {
+          const u = uri.parse(url, true);
+
           expect(request.session['oauth2:www.example.com'].state).to.have.length(32);
           expect(request.session['oauth2:www.example.com'].state).to.equal(u.query.state);
         });
-        
-        it('should preserve other data in session', function() {
+
+        it('should preserve other data in session', () => {
           expect(request.session['oauth2:www.example.com'].foo).to.equal('bar');
         });
       }); // that redirects to service provider with other data in session
-      
-      describe('that errors due to lack of session support in app', function() {
-        var request, err;
-  
-        before(function(done) {
+
+      describe('that errors due to lack of session support in app', () => {
+        let err;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .error(function(e) {
+            .error((e) => {
               err = e;
               done();
             })
-            .req(function(req) {
-              request = req;
+            .req(() => {
             })
             .authenticate();
         });
-  
-        it('should error', function() {
-          expect(err).to.be.an.instanceof(Error)
+
+        it('should error', () => {
+          expect(err).to.be.an.instanceof(Error);
           expect(err.message).to.equal('OAuth 2.0 authentication requires session support when using state. Did you forget to use express-session middleware?');
         });
       }); // that errors due to lack of session support in app
-      
     }); // issuing authorization request
-    
-    describe('issuing authorization request to authorization server using authorization endpoint that has query parameters including state', function() {
-      var strategy = new OAuth2Strategy({
+
+    describe('issuing authorization request to authorization server using authorization endpoint that has query parameters including state', () => {
+      const strategy = new OAuth2Strategy({
         authorizationURL: 'https://www.example.com/oauth2/authorize?foo=bar&state=baz',
         tokenURL: 'https://www.example.com/oauth2/token',
         clientID: 'ABC123',
@@ -118,43 +114,44 @@ describe('OAuth2Strategy', function() {
         callbackURL: 'https://www.example.net/auth/example/callback',
         state: true
       },
-      function(accessToken, refreshToken, profile, done) {});
-      
-      
-      describe('that redirects to service provider', function() {
-        var request, url;
-  
-        before(function(done) {
+      (() => {}));
+
+
+      describe('that redirects to service provider', () => {
+        let request; let
+          url;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .redirect(function(u) {
+            .redirect((u) => {
               url = u;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
               req.session = {};
             })
             .authenticate();
         });
-  
-        it('should be redirected', function() {
-          var u = uri.parse(url, true);
+
+        it('should be redirected', () => {
+          const u = uri.parse(url, true);
           expect(u.query.foo).equal('bar');
           expect(u.query.state).to.have.length(32);
         });
-      
-        it('should save state in session', function() {
-          var u = uri.parse(url, true);
-        
+
+        it('should save state in session', () => {
+          const u = uri.parse(url, true);
+
           expect(request.session['oauth2:www.example.com'].state).to.have.length(32);
           expect(request.session['oauth2:www.example.com'].state).to.equal(u.query.state);
         });
       }); // that redirects to service provider
-      
-    }); // issuing authorization request to authorization server using authorization endpoint that has query parameters including state
-    
-    describe('processing response to authorization request', function() {
-      var strategy = new OAuth2Strategy({
+    }); /* issuing authorization request to authorization server using authorization
+      endpoint that has query parameters including state */
+
+    describe('processing response to authorization request', () => {
+      const strategy = new OAuth2Strategy({
         authorizationURL: 'https://www.example.com/oauth2/authorize',
         tokenURL: 'https://www.example.com/oauth2/token',
         clientID: 'ABC123',
@@ -162,158 +159,169 @@ describe('OAuth2Strategy', function() {
         callbackURL: 'https://www.example.net/auth/example/callback',
         state: true
       },
-      function(accessToken, refreshToken, profile, done) {
+      ((accessToken, refreshToken, profile, done) => {
         if (accessToken !== '2YotnFZFEjr1zCsicMWpAA') { return done(new Error('incorrect accessToken argument')); }
         if (refreshToken !== 'tGzv3JOkF0XG5Qx2TlKWIA') { return done(new Error('incorrect refreshToken argument')); }
         if (typeof profile !== 'object') { return done(new Error('incorrect profile argument')); }
         if (Object.keys(profile).length !== 0) { return done(new Error('incorrect profile argument')); }
 
         return done(null, { id: '1234' }, { message: 'Hello' });
-      });
+      }));
 
-      strategy._oauth2.getOAuthAccessToken = function(code, options, callback) {
+      strategy._oauth2.getOAuthAccessToken = function getOAuthAccessToken(code, options, callback) {
         if (code !== 'SplxlOBeZQQYbYS6WxSbIA') { return callback(new Error('incorrect code argument')); }
         if (options.grant_type !== 'authorization_code') { return callback(new Error('incorrect options.grant_type argument')); }
         if (options.redirect_uri !== 'https://www.example.net/auth/example/callback') { return callback(new Error('incorrect options.redirect_uri argument')); }
 
         return callback(null, '2YotnFZFEjr1zCsicMWpAA', 'tGzv3JOkF0XG5Qx2TlKWIA', { token_type: 'example' });
-      }
-      
-      
-      describe('that was approved', function() {
-        var request
-          , user
-          , info;
-  
-        before(function(done) {
+      };
+
+
+      describe('that was approved', () => {
+        let request;
+
+
+        let user;
+
+
+        let info;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .success(function(u, i) {
+            .success((u, i) => {
               user = u;
               info = i;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
-            
+
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
               req.session = {};
               req.session['oauth2:www.example.com'] = {};
-              req.session['oauth2:www.example.com']['state'] = 'DkbychwKu8kBaJoLE5yeR5NK';
+              req.session['oauth2:www.example.com'].state = 'DkbychwKu8kBaJoLE5yeR5NK';
             })
             .authenticate();
         });
-  
-        it('should supply user', function() {
+
+        it('should supply user', () => {
           expect(user).to.be.an('object');
           expect(user.id).to.equal('1234');
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Hello');
         });
-      
-        it('should remove state from session', function() {
+
+        it('should remove state from session', () => {
           expect(request.session['oauth2:www.example.com']).to.be.undefined;
         });
       }); // that was approved
-      
-      describe('that was approved with other data in the session', function() {
-        var request
-          , user
-          , info;
-  
-        before(function(done) {
+
+      describe('that was approved with other data in the session', () => {
+        let request;
+
+
+        let user;
+
+
+        let info;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .success(function(u, i) {
+            .success((u, i) => {
               user = u;
               info = i;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
-            
+
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
               req.session = {};
               req.session['oauth2:www.example.com'] = {};
-              req.session['oauth2:www.example.com']['state'] = 'DkbychwKu8kBaJoLE5yeR5NK';
+              req.session['oauth2:www.example.com'].state = 'DkbychwKu8kBaJoLE5yeR5NK';
               req.session['oauth2:www.example.com'].foo = 'bar';
             })
             .authenticate();
         });
-  
-        it('should supply user', function() {
+
+        it('should supply user', () => {
           expect(user).to.be.an('object');
           expect(user.id).to.equal('1234');
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Hello');
         });
-      
-        it('should preserve other data from session', function() {
+
+        it('should preserve other data from session', () => {
           expect(request.session['oauth2:www.example.com'].state).to.be.undefined;
           expect(request.session['oauth2:www.example.com'].foo).to.equal('bar');
         });
       }); // that was approved with other data in the session
-      
-      describe('that fails due to state being invalid', function() {
-        var request
-          , info, status;
-  
-        before(function(done) {
+
+      describe('that fails due to state being invalid', () => {
+        let request;
+
+
+        let info;
+
+
+        let status;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .fail(function(i, s) {
+            .fail((i, s) => {
               info = i;
               status = s;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
-            
+
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK-WRONG';
               req.session = {};
               req.session['oauth2:www.example.com'] = {};
-              req.session['oauth2:www.example.com']['state'] = 'DkbychwKu8kBaJoLE5yeR5NK';
+              req.session['oauth2:www.example.com'].state = 'DkbychwKu8kBaJoLE5yeR5NK';
             })
             .authenticate();
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Invalid authorization request state.');
         });
-      
-        it('should supply status', function() {
+
+        it('should supply status', () => {
           expect(status).to.equal(403);
         });
-      
-        it('should remove state from session', function() {
+
+        it('should remove state from session', () => {
           expect(request.session['oauth2:www.example.com']).to.be.undefined;
         });
       }); // that fails due to state being invalid
-      
-      describe('that fails due to provider-specific state not found in session', function() {
-        var request
-          , info, status;
-  
-        before(function(done) {
+
+      describe('that fails due to provider-specific state not found in session', () => {
+        let info;
+        let status;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .fail(function(i, s) {
+            .fail((i, s) => {
               info = i;
               status = s;
               done();
             })
-            .req(function(req) {
-              request = req;
-            
+            .req((req) => {
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
@@ -321,31 +329,29 @@ describe('OAuth2Strategy', function() {
             })
             .authenticate();
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Unable to verify authorization request state.');
         });
-      
-        it('should supply status', function() {
+
+        it('should supply status', () => {
           expect(status).to.equal(403);
         });
       }); // that fails due to state not found in session
-      
-      describe('that fails due to provider-specific state lacking state value', function() {
-        var request
-          , info, status;
-  
-        before(function(done) {
+
+      describe('that fails due to provider-specific state lacking state value', () => {
+        let info;
+        let status;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .fail(function(i, s) {
+            .fail((i, s) => {
               info = i;
               status = s;
               done();
             })
-            .req(function(req) {
-              request = req;
-            
+            .req((req) => {
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
@@ -354,50 +360,45 @@ describe('OAuth2Strategy', function() {
             })
             .authenticate();
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Unable to verify authorization request state.');
         });
-      
-        it('should supply status', function() {
+
+        it('should supply status', () => {
           expect(status).to.equal(403);
         });
       }); // that fails due to provider-specific state lacking state value
-      
-      describe('that errors due to lack of session support in app', function() {
-        var request
-          , err;
-  
-        before(function(done) {
+
+      describe('that errors due to lack of session support in app', () => {
+        let err;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .error(function(e) {
+            .error((e) => {
               err = e;
               done();
             })
-            .req(function(req) {
-              request = req;
-            
+            .req((req) => {
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
             })
             .authenticate();
         });
-  
-        it('should error', function() {
-          expect(err).to.be.an.instanceof(Error)
+
+        it('should error', () => {
+          expect(err).to.be.an.instanceof(Error);
           expect(err.message).to.equal('OAuth 2.0 authentication requires session support when using state. Did you forget to use express-session middleware?');
         });
       }); // that errors due to lack of session support in app
-      
     }); // processing response to authorization request
-    
   }); // using default session state store
-  
-  
-  describe('using default session state store with session key option', function() {
-    var strategy = new OAuth2Strategy({
+
+
+  describe('using default session state store with session key option', () => {
+    const strategy = new OAuth2Strategy({
       authorizationURL: 'https://www.example.com/oauth2/authorize',
       tokenURL: 'https://www.example.com/oauth2/token',
       clientID: 'ABC123',
@@ -406,101 +407,100 @@ describe('OAuth2Strategy', function() {
       state: true,
       sessionKey: 'oauth2:example'
     },
-    function(accessToken, refreshToken, profile, done) {
+    ((accessToken, refreshToken, profile, done) => {
       if (accessToken !== '2YotnFZFEjr1zCsicMWpAA') { return done(new Error('incorrect accessToken argument')); }
       if (refreshToken !== 'tGzv3JOkF0XG5Qx2TlKWIA') { return done(new Error('incorrect refreshToken argument')); }
       if (typeof profile !== 'object') { return done(new Error('incorrect profile argument')); }
       if (Object.keys(profile).length !== 0) { return done(new Error('incorrect profile argument')); }
 
       return done(null, { id: '1234' }, { message: 'Hello' });
-    });
+    }));
 
-    strategy._oauth2.getOAuthAccessToken = function(code, options, callback) {
+    strategy._oauth2.getOAuthAccessToken = function getOAuthAccessToken(code, options, callback) {
       if (code !== 'SplxlOBeZQQYbYS6WxSbIA') { return callback(new Error('incorrect code argument')); }
       if (options.grant_type !== 'authorization_code') { return callback(new Error('incorrect options.grant_type argument')); }
       if (options.redirect_uri !== 'https://www.example.net/auth/example/callback') { return callback(new Error('incorrect options.redirect_uri argument')); }
 
       return callback(null, '2YotnFZFEjr1zCsicMWpAA', 'tGzv3JOkF0XG5Qx2TlKWIA', { token_type: 'example' });
-    }
-    
-    
-    describe('issuing authorization request', function() {
-      
-      describe('that redirects to service provider', function() {
-        var request, url;
-  
-        before(function(done) {
+    };
+
+
+    describe('issuing authorization request', () => {
+      describe('that redirects to service provider', () => {
+        let request; let
+          url;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .redirect(function(u) {
+            .redirect((u) => {
               url = u;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
               req.session = {};
             })
             .authenticate();
         });
-  
-        it('should be redirected', function() {
-          var u = uri.parse(url, true);
+
+        it('should be redirected', () => {
+          const u = uri.parse(url, true);
           expect(u.query.state).to.have.length(32);
         });
-      
-        it('should save state in session', function() {
-          var u = uri.parse(url, true);
-        
+
+        it('should save state in session', () => {
+          const u = uri.parse(url, true);
+
           expect(request.session['oauth2:example'].state).to.have.length(32);
           expect(request.session['oauth2:example'].state).to.equal(u.query.state);
         });
       }); // that redirects to service provider
-      
     }); // issuing authorization request
-    
-    describe('processing response to authorization request', function() {
-      
-      describe('that was approved', function() {
-        var request
-          , user
-          , info;
-  
-        before(function(done) {
+
+    describe('processing response to authorization request', () => {
+      describe('that was approved', () => {
+        let request;
+
+
+        let user;
+
+
+        let info;
+
+        before((done) => {
           chai.passport.use(strategy)
-            .success(function(u, i) {
+            .success((u, i) => {
               user = u;
               info = i;
               done();
             })
-            .req(function(req) {
+            .req((req) => {
               request = req;
-            
+
               req.query = {};
               req.query.code = 'SplxlOBeZQQYbYS6WxSbIA';
               req.query.state = 'DkbychwKu8kBaJoLE5yeR5NK';
               req.session = {};
               req.session['oauth2:example'] = {};
-              req.session['oauth2:example']['state'] = 'DkbychwKu8kBaJoLE5yeR5NK';
+              req.session['oauth2:example'].state = 'DkbychwKu8kBaJoLE5yeR5NK';
             })
             .authenticate();
         });
-  
-        it('should supply user', function() {
+
+        it('should supply user', () => {
           expect(user).to.be.an('object');
           expect(user.id).to.equal('1234');
         });
-  
-        it('should supply info', function() {
+
+        it('should supply info', () => {
           expect(info).to.be.an('object');
           expect(info.message).to.equal('Hello');
         });
-      
-        it('should remove state from session', function() {
+
+        it('should remove state from session', () => {
           expect(request.session['oauth2:example']).to.be.undefined;
         });
       }); // that was approved
-      
     }); // processing response to authorization request
-    
   }); // using default session state store with session key option
-  
 });